WIP: Items Checked Out
Discovered that to call open-ils.actor.user.checked_out you need
the userID, and for that you need to call open-ils.auth.session.retrieve,
and for that you need the IDL.  So this bit will take a few revs.

Also make Account a class so it has reference semantics; else when
you pass it as a parameter you get a copy.