First working version of Search > Results segue
Had to totally recreate the storyboard using Editor > Refactor to Storyboard
or else I kept getting the error "Search Results View Controller" is unreachable
because it has no entry points.  Which was bogus because it did have an entry point.
My best conclusion is that xcode got confused and to fix it I had to use xcode
to build everything.