Acorn: Handle online items even in cases where item_form is wrong
Some records, e.g. https://acorn.biblio.org/eg/opac/record/2891957
do not have item_form="o", so as backup we wait for the marcxml to
load and treat as an online item if the item has links.

We call setupButtons again after all promises resolve, which gives
us a chance to catch the rogue online item and change the buttons.
In that rare case the buttons flash.  We clear the button targets
now before adding, because otherwise they stack and you get both.