Fix bug where Place Hold button scrolls off screen
On a small screen like iPhone 5s or SE, transitioning the expiration
date picker into view can force the Place Hold button off screen.
This workaround is not perfect; it scrolls too high and if you
manually scroll again it snaps back down. But it is Good Enough
and I was not able to do better.