Fix Show Card for dark mode
Because some dark mode logos are transparent with light foreground,
they do not show up against a white background.  So for the logo to
show up the background has to be adaptive.  Making the background
adaptive also allows the barcode text and warning (if any) to be
legible, without (I hope) sacrificing anything from the barcode,
which has built in quiet zones on either side.