Fix hemlock-pm#17 - ILL checkouts have no title
Do not fetch MRA if doc_id == -1, but do fetch the asset copy (acp).
The way I implemented this is by adding an extra stage to the promise
chain, and returning an empty promise (or a promise of an empty
OSRFObject) as a conditional.  I don't like it that much but it
does work.