Fix bug where patron without day_phone could not place hold
Because the hold UX is not visible for NOBLE, and because the
default notify is email:phone, the Switch was turned on, but
there was no phone number present and a hold attempt failed with
"Phone number field cannot be empty".  Instead, only turn on
the Switch when we have a default=true and a value.  This goes
for the SMS switch as well.