Ignore error loading metadata for holds
The error (expected object, got empty) is not useful for the user,
and it seems the OPAC squelches such issues too.  In this case,
one item on hold is deleted.  In order to detect a deleted record,
it seems you need to fetch the BRE and check the deleted attribute
(see 42e32196856bb085d4e3c4244f66483fd4fb6e55).