Tweak layout of search results
but async loading of a big list of long titles is
still a bit janky.  If you scroll fast to a title
that wraps, the cell height might have been estimated
with an empty string that would only take 1 line.
Bringing the cell off-screen and back on-screen fixes
it but the scrolling hitches.