Improve UX of date pickers in Hold VC
Because ahr objects can have null dates but UIDatePickers cannot,
we make certain adjustments:
* dim the controls when disabling them
* display reasonable dates (now + 1 year)
* don't actually set a thaw date unless the user changes the picker

Allowing a null thaw date is convenient; it's suspended until you
pick a date.  Always showing a thaw date of now + 1 year is a
compromise that is fixed by allowing the user to pick a specific date.