Loosen coupling between Login and Main VC
The goal is to allow Acorn to launch a different Main VC and
we can't do that if Login has a hard reference (named segue).
Now the Login and Main storyboards each have their respective
NavigationControllers, and the transition is done programmatically
via popToMain() or popToLogin().  This is much better.

Slightly worse is that the transition from Login to Main is no
longer animated.