Cross-fade between Login and Main
Animate the transition of switching between the 2 root VCs.
BTW, 2 root VCs are useful because the main one has a title and
largeTitles=true.  Note that the initialVC on both of them is
actually the UINavigationController, so that if you want the real
VC you have to look at its topViewController.

Other related changes:
* Set LoginVC.isAddingAccount=true and if so use that var to
  avoid auto-login.  Do not clear lastUsedCredential, because we'll
  need that to implement a cancel action.  Currently if you start to
  add an account but change your mind, all you can do is kill the app.
* Remove never-used var and useless overrides from LoginVC.
* Clean up loggin of creds and LoginVC.