Make acorn use old UX on iOS 12
There is only 1 iOS 12 device running Acorn (and it
might be me :) but this prevents a crash.  At some
point in the near future I will drop support for
iOS 12 on all apps, but it would be a headache to
do it only for some.  Setting min=14.0 just on acorn causes
MD5 deprecation warnings but the replacement is not
available until 13.0.