Improve "The operation couldn't be completed" error message
That is the default localizedDescription for Error objects
that do not implement the LocalizedError protocol.  Also, log
an error with Analytics when OSRF decode errors happen.
See github.com/mxcl/LegibleError for a fuller description
of the localizedDescription issue and a more general fix.
At this time I do not want to add another dependency.