Fix missing title bug when paging through results
If you got a lot of results, but they were not all prefetched,
then you could page one-by-one to a record with missing title.
Fix it the easiest way instead of rethinking the whole
MBRecord vs AsyncRecord thing.  So for now AsyncRecord is only
used in ResultsVC.  Also prevent refetching MARC when paging
back.