fix: Update org contact details on every Library Info open
Fix it so you can change the org email and org phone number in the
database and it will show up the next time you load the Library Info
screen, instead of when the cache expires.

We intended to refresh the org contact details (stuff on the aou object)
when Library Info was loaded, but some async mistakes broke it.
The "when ready" was called too soon and too generally.  Now, each
async load invokes its own "on X loaded" func and updates the UX.