Make loadHoldQueueStats run in parallel task
and note that async-let creates a Task.  The doc speaks of async-let
as if the constant is of the type returned by the async function,
but really I think it is more correct to think of it as type Task,
because you have to do assign new constants with `await` to get
at the values, e.g.

  async let xTask = callAsyncFunction()
  async let yTask = anotherAsyncFunc()
  let (xVal, yVal) = try await (xTask, yTask)