Fix table header incorrectly showing "no results"
At the time tableView(:titleforHeaderInSection:) is called,
the activityIndicator is not animating, and so the header
was showing "No results" until the search finished.
Now displays "Searching..." until didCompleteSearch is true,
and make sure it gets set even in error conditions.  Remember
that in commit ee915a62 it seems "Searching..." got stuck on.