Fix: Filled buttons have white outline in dark mode
Now that min iOS is 15, we can use UIButton.Configuration.  Use the
builtin styles instead of rolling our own, and set a global accent color.

This has a number of benefits:
* filled style does not have a weird looking white border when enabled,
  and disabled state behaves well
* gray style looks better than our outline
* titles use dynamic type
* have the option to use an in-button activity indicator (not yet)
* have the option to set styles using Interface Builder (not yet)
* other controls are automatically tinted, like the checkmark

To make all buttons consistent, in IB I changed the few Buttons that had
a style to Style: Default and set the config in code.