Fix tests after grand rename in 7fc4b755
That commit renamed all targets and that caused some of the
values of Product Name to change.  It also changed the name
of the Hemlock module and Hemlock.app, breaking the tests.
Keep the lowercase scheme names, and manually override the
Product Name back to the old name for some schemes
(Hemlock, OWWL, PINES, NOBLE).

NB: Xcode fails to make case-only renames, so if you want to
rename a scheme and change only the case, you have to perform
a 2-stage rename.